enhance(sdk): remove the `fuzzyArn`, instead use the lazy value from pulumi
Before, the type used to construct the permission list was `aws.types.input.iam.GetPolicyDocumentStatement` type, it only allows passing strings as arguments, and we used `fuzzyArn` to build the ARN through string concatenation. This approach is messy and unsafe. Now, we've switched the permission type to `aws.types.input.iam.GetPolicyDocumentStatementArgs`, which allows us to pass lazy values as arguments. This means we can configure the permission using the ARN from the Pulumi resource.