feat: create k3d container image to streamline Kubernetes experience with Pluto
This commit addresses a specific requirement from Pluto for Kubernetes. Previously, users needed to manually install k3d, kubectl, knative, and other tools before they could try Pluto. This process created a gap in user experience.

To bridge this gap, this commit introduces a container image that includes docker-cli, k3d, kubectl, and scripts to create the k3s cluster and install knative. The aim is to simplify the user experience when trying Pluto in a Kubernetes environment.