feat(sdk): optimize cold-start performance
Move import statements to container initialization to enhance runtime performance.

By shifting import statements to container initialization rather than at invocation time, we ensure that the import process is completed with more resources available, leading to improved cold-start performance.

For the `hello` and `store` path handlers in the quickstart example, this change reduces the latency of the first request from 9s to 2.4s. Increasing the memory size from 128MB to 1024MB further cuts the first request latency down to 1.8s. These measurements were taken in China, and the lambda instance is deployed in the us-east-1 region.